resolved sonar issue for SubutaiExecutors, according to sonar issue description ->
This Object.finalize() method is called by the garbage collector on an object when garbage collection determines that there are no more references to the object. But there is absolutely no warranty that this method will be called AS SOON AS the last references to the object are removed. It can be few microseconds to few minutes later. So when some system resources need to be disposed by an object it's better to not rely on this asynchronous mechanism to dispose them.

so with a help of http://howtodoinjava.com/2012/10/31/why-not-to-use-finalize-method-in-java/
I have applied described approach