resolved sonar issue for PluginManagerImpl issue groups resolved in this commit removed some redundancies ******** Throwable.printStackTrace(...) should never be called
Throwable.printStackTrace(...) prints a throwable and its stack trace to some stream.

Loggers should be used instead to print throwables, as they have many advantages:

    Users are able to easily retrieve the logs.
    The format of log messages is uniform and allow users to browse the logs easily.

********

Correctness - Method call passes null for nonnull parameter

This method call passes a null value for a nonnull method parameter. Either the parameter is annotated as a parameter that should always be nonnull, or analysis has shown that it will always be dereferenced.