sonar fix Lambdas and anonymous classes should not have too many lines
Anonymous classes and lambdas (with Java 8) are a very convenient and compact way to inject a behavior without having to create a dedicated class. But those anonymous inner classes and lambdas should be used only if the behavior to be injected can be defined in a few lines of code, otherwise the source code can quickly become unreadable.

Avoid commented-out lines of code

Here are the main reasons why commented code is a code smell :
It always raises more questions than it gives answers
Everybody will forget very quickly how relevant the commented code is
This is distraction when going down the code as it stops the flow of eyes
It is a bad SCM engine : Subversion, CVS and Git are really more trustworthy !
The simple fact of understanding why code was commented out in the first place can take a lot of time