sonar issue fix for ResourceHostEntity Loggers should be "private static final" and should share a naming convention
Whatever is the logging framework (logback, log4j, commons-logging, java.util.logging, ...), a logger should be:
private: Never be accessible outside of its parent class. If another class needs to log something, it should instantiate its own logger.
static: Not dependent on an instance of a class (an object). When logging something, contextual information can of course be provided in the messages but the logger should be created at class level to prevent creating a logger along with each object.
final: Be created once and only once per class.

The members of an interface declaration or class should appear in a pre-defined order

According to the Java Code Conventions as defined by Oracle, the parts of a class or interface declaration should appear in the following order in the source files:
Class and instance variables
Constructors
Methods