configured filedb to be osgi compatible
For reference:
OSGi does not allow to get classloader of another bundle thus some
classes from another bundle will not be accessible due to bundle
restrictions, to get there you will have to make static reference in bundle
Manifest file or dynamically import, which is not recommended way to
build osgi bundles.

So as a workaround and make life easier without biting own tail and walking
on the same place it is much more better to specifically pass required classloader and reuse it where it is needed.